requirements.txt
django-waffle
djangorestframework~=3.12.1
mysqlclient
djangorestframework-simplejwt
pipenv
gunicorn
django_rest_passwordreset
Pillow==6.2.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-cleanup==4.0.0
django-crispy-forms==1.8.0
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
django-currentuser==0.5.1
django-livesync== 0.5
numpy
django-debug-toolbar
django-oauth-toolkit
djangorestframework-oauth
djangorestframework_simplejwt
djoser~=2.1.0
django-filter==2.4.0
django-cors-headers
django-rest-auth==0.9.5
django-silk==4.1.0
sorl-thumbnail==12.7.0
django-report-builder==6.4.0
django-admin-interface==0.14.1
django-otp==1.0.2
qrcode==6.1
django-sass==1.0.0
graphene~=2.1.9
django-filter>=2
flask-login~=0.5.0
django-graphql-jwt~=0.3.4